refactor: Extract duplicate functions to shared tasks package
Address nitpick comments from PR review:
- Implement fail-fast error handling in cmd/status.go (return immediately on errors instead of continuing)
- Remove duplicate code by creating shared internal/tasks/helpers.go package
- Move FilterTasksByStatus, SortTasksByPriority, GenerateTaskID, and CalculateTaskStats to shared package
- Update both cmd/status.go and internal/tui/dashboard.go to use shared functions
- Remove duplicate function definitions and type declarations
